chore: Move pytest fixtures to proper `conftest.py`
