+++ tests/unit/test_raster_math.py
@@ -1,59 +1,6 @@
 import numpy as np
-import pytest
 import rioxarray
-import xarray as xr
-from affine import Affine
 
-
-# --- Fixtures ---
-@pytest.fixture
-def mock_raster_factory(tmp_path):
-    """
-    Returns a factory function that creates a GeoTIFF in a temporary directory.
-
-    Arguments:
-    - filename: Name of the file to create.
-    - data_array: Numpy array of values.
-    - origin: Tuple (min_x, max_y) specifying the top-left corner coordinates.
-    - res: Pixel resolution (default 1.0).
-    - crs: Coordinate reference system (default EPSG:4326).
-    """
-    from worldpoppy.config import WGS84_CRS
-
-    def _create(
-        filename, data_array, origin=(0, 0), res=1.0, crs=WGS84_CRS, nodata=-9999
-    ):
-        height, width = data_array.shape
-        x_start, y_start = origin
-
-        # 1. Create the transform automatically
-        # (Top-left X, Pixel Width, 0, Top-left Y, 0, -Pixel Height)
-        transform = Affine.translation(x_start, y_start) * Affine.scale(res, -res)
-
-        # 2. Generate Coordinate Arrays (Pixel Centers)
-        # X: start + half_res, start + 1.5*res, ...
-        xs = np.arange(width) * res + x_start + (res / 2)
-        # Y: start - half_res, start - 1.5*res, ... (negative because Y goes down)
-        ys = y_start - (np.arange(height) * res) - (res / 2)
-
-        # 3. Build DataArray
-        da = xr.DataArray(data_array, dims=("y", "x"), coords={"y": ys, "x": xs})
-
-        # 4. Write Geospatial Metadata
-        da.rio.write_crs(crs, inplace=True)
-        da.rio.write_transform(transform, inplace=True)
-        da.rio.write_nodata(nodata, inplace=True)
-
-        # 5. Save
-        file_path = tmp_path / filename
-        da.rio.to_raster(file_path)
-
-        return file_path
-
-    return _create
-
-
-# --- Tests ---
 
 def test_merge_preserves_count(mock_raster_factory):
     """
